Modify code to use A-Star 32U4 SV board
Convert code to use A-Star 32U4 SV board, for which there
are several differences.
* 5 - 36 V input is possible, and BATLEV divider is different
* No IMU built in
* No encoders built-in
* MAX14870 motor drivers are 5 - 36 V

Call readBatteryMillivoltsSV instead of readBatteryMilliVolts, due to
the difference in voltage dividers. Detect the power source between USB,
6x AA cells, and 6x 14500 (3.7V) cells, based upon initial voltage, and
use that to determine a low-voltage threshold. This prevents the buzzer
from buzzing the entire time if using a computer to flash the A-Star32U4
via USB also. When using 6x 14500 cells, should probably also pull the
RPISHNDN pin in addition to turning on the buzzer when the LV threshold
is reached.

Care should be taken to limit power to the motors if using the simple Romi
motors.

Ideally, one could detect which type of Pololu 32U4 board was connected and
have the FW adapt to the differences, or load different FW based on type.
Significantly different hardware configurations also likely require
soldering and a good deal of FW customization.